Removed obsolete ModelCreator
